Decouple daemon and container from the stats collector.
